+++ CHANGELOG.md
@@ -4,7 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 ## Added
-- Free draw
+- Free draw (#1)
+- Drawing style history (#4)
 
 ## [1.0.0] - 2016-02-28
 First version.